requirements_vit_v1.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
fastapi==0.109.2
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2024.2.0
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.20.3
idna==3.6
ipykernel==6.29.2
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.15.0
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
PyMuPDF==1.23.25
PyMuPDFb==1.23.22
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tokenizers==0.15.2
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
